make math::bit* functions receive a uint32_t

+++ include/util/Math.h
@@ -4,15 +4,15 @@
 #include <cstdint>
 
 namespace gameboy::math {
-    inline bool bit_test(uint8_t n, uint8_t bit) {
+    inline bool bit_test(uint32_t n, uint32_t bit) {
         return ((n >> bit) & 0x1u) == 0x1u;
     }
 
-    inline void bit_set(uint8_t& n, uint8_t bit) {
+    inline void bit_set(uint32_t& n, uint32_t bit) {
         n |= 0x1u << bit;
     }
 
-    inline void bit_reset(uint8_t& n, uint8_t bit) {
+    inline void bit_reset(uint32_t& n, uint32_t bit) {
         n &= ~(0x1u << bit);
     }
 }
